refactor: Update configuration imports and enhance dependency management
- Moved configuration imports from gemini_sre_agent.config to gemini_sre_agent.legacy_config for better organization.
- Added new dependencies for google-generativeai and updated existing ones in pyproject.toml and uv.lock.
- Removed obsolete config.py file to streamline the configuration system.
- Enhanced test coverage for configuration management utilities and ensured all tests are passing.
