+++ pyproject.toml
@@ -25,6 +25,7 @@ dependencies = [
   "google-auth>=2.29.0",
   "google-auth-oauthlib>=1.2.0",
   "google-auth-httplib2>=0.2.0",
+  "google-generativeai>=0.3.0",
 ]
 
 [tool.pytest.ini_options]
